@@ -1,6 +1,6 @@
 {
   "name": "react-adsense",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "A Simple React component for Google AdSense and Baidu advertisement.",
   "main": "lib/index.js",
   "scripts": {
@@ -32,29 +32,30 @@
   },
   "homepage": "https://github.com/hustcc/react-adsense",
   "devDependencies": {
-    "babel-cli": "^6.6.0",
-    "babel-core": "^6.3.15",
-    "babel-eslint": "^5.0.0-beta4",
-    "babel-loader": "^6.2.0",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-react-hmre": "^1.0.0",
-    "eslint": "^1.10.3",
-    "eslint-plugin-babel": "^3.0.0",
-    "eslint-plugin-react": "^3.11.3",
-    "react": "^0.14.0",
-    "react-dom": "^0.14.0",
-    "webpack": "^1.12.9",
-    "css-loader": "~0.16.0",
-    "jsx-loader": "~0.13.2",
-    "style-loader": "~0.12.3",
-    "file-loader": "~0.8.4",
-    "rimraf": "^2.5.0",
-    "cross-env": "^1.0.8"
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-eslint": "^7.2.3",
+    "babel-loader": "^7.0.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-react-hmre": "^1.1.1",
+    "cross-env": "^5.0.0",
+    "css-loader": "^0.28.4",
+    "eslint": "^3.19.0",
+    "eslint-plugin-babel": "^4.1.1",
+    "eslint-plugin-react": "^7.0.1",
+    "file-loader": "^0.11.1",
+    "jsx-loader": "^0.13.2",
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "rimraf": "^2.6.1",
+    "style-loader": "^0.18.1",
+    "webpack": "^2.6.1"
   },
-  "dependencies": {
-  },
+  "dependencies": {},
   "peerDependencies": {
-    "react": ">=0.13.2 || ^0.14 || ^15.0.0"
+    "react": ">=0.13.2 || ^0.14 || ^15.0.0 || >=16.0.0-alpha.1 <17.0.0",
+    "prop-types": ">=15.0.0"
   }
 }
