@@ -1,39 +1,66 @@
 'use strict';
 
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
+
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var Google = _react2.default.createClass({
-  displayName: 'Google',
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
-  propTypes: {
-    style: _react2.default.PropTypes.object,
-    client: _react2.default.PropTypes.string.isRequired,
-    slot: _react2.default.PropTypes.string.isRequired,
-    format: _react2.default.PropTypes.string
-  },
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-  componentDidMount: function componentDidMount() {
-    if (window) (window.adsbygoogle = window.adsbygoogle || []).push({});
-  },
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
-  getDefaultProps: function getDefaultProps() {
-    return {
-      style: { display: 'block' },
-      format: 'auto'
-    };
-  },
+var Google = function (_React$Component) {
+  _inherits(Google, _React$Component);
 
-  render: function render() {
-    return _react2.default.createElement('ins', { className: 'adsbygoogle',
-      style: this.props.style,
-      'data-ad-client': this.props.client,
-      'data-ad-slot': this.props.slot,
-      'data-ad-format': this.props.format });
+  function Google() {
+    _classCallCheck(this, Google);
+
+    return _possibleConstructorReturn(this, (Google.__proto__ || Object.getPrototypeOf(Google)).apply(this, arguments));
   }
-});
 
-module.exports = Google;
+  _createClass(Google, [{
+    key: 'componentDidMount',
+    value: function componentDidMount() {
+      if (window) (window.adsbygoogle = window.adsbygoogle || []).push({});
+    }
+  }, {
+    key: 'render',
+    value: function render() {
+      return _react2.default.createElement('ins', { className: 'adsbygoogle',
+        style: this.props.style,
+        'data-ad-client': this.props.client,
+        'data-ad-slot': this.props.slot,
+        'data-ad-format': this.props.format });
+    }
+  }]);
+
+  return Google;
+}(_react2.default.Component);
+
+exports.default = Google;
+;
+
+Google.propTypes = {
+  style: _react2.default.PropTypes.object, // eslint-disable-line
+  client: _react2.default.PropTypes.string.isRequired,
+  slot: _react2.default.PropTypes.string.isRequired,
+  format: _react2.default.PropTypes.string
+};
+
+Google.defaultProps = {
+  style: { display: 'block' },
+  format: 'auto'
+};
