@@ -14,8 +14,9 @@ var _baidu2 = _interopRequireDefault(_baidu);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var AdSense = {};
-AdSense.Google = _google2.default;
-AdSense.Baidu = _baidu2.default;
+var AdSense = {
+  Google: _google2.default,
+  Baidu: _baidu2.default
+};
 
 exports.default = AdSense;
