@@ -1,8 +1,9 @@
 import Google from './google';
 import Baidu from './baidu';
 
-let AdSense = {};
-AdSense.Google = Google;
-AdSense.Baidu = Baidu;
+const AdSense = {
+  Google,
+  Baidu
+};
 
 export default AdSense;
